MOPIA_ProClean1-main/requirements.txt
Django==5.2
#psycopg2==2.9.6
psycopg2-binary==2.9.6
